Created a selection cursor script for possible menu usage. Fixed issue where two colliding boxes get stuck together. Main Menu uses the InputManager for mouse and keyboard control. Pause menu uses time scale to pause game. GameManager correctly resets to default values for file info when game goes back to main menu. Weapon inventory items temporarily go into first quick slot, and item can activate attack animation.
In order to play the correct animations, it all depends on the weapon type of the item that the player is using in the corresponding quick slot. Then the corresponding weapon's animation should play at the same time. Probably should use a separate layer for weapon animations and not a completely separate animator.